* tests/socket.c (SHOW_RESULT): Add a cast to avoid compiler         warnings about the format string not matching the parameters.         * tests/testfs.c: rename strcmp and strcpy to mystrcmp and         mystrcpy to avoid clashes with the normal implementations.

+++ packages/io/fileio/current/tests/socket.c
@@ -101,7 +101,7 @@
 //--------------------------------------------------------------------------
 
 #define SHOW_RESULT( _fn, _res ) \
-diag_printf(#_fn " returned %d %s\n", _res, _res<0?strerror(errno):"");
+diag_printf(#_fn " returned %d %s\n", (int) _res, _res<0?strerror(errno):"");
 
 //--------------------------------------------------------------------------
 // Thread stack.
